Chore: Moves helper functions to helper.py
Feat: Adds user_not_logged_in function decorator
